Restructures model tests into seperate submodules
